Extract generic web preview verification dispatch
